requirements.txt
altgraph==0.17.4
async-timeout==5.0.1
blessed==1.21.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
dnspython==2.7.0
editor==1.6.6
idna==3.10
ifaddr==0.2.0
inquirer==3.4.0
joblib==1.5.0
macholib==1.16.3
numpy==2.2.6
packaging==25.0
pandas==2.2.3
polars==1.29.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
pymongo==4.13.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
readchar==4.2.1
requests==2.32.3
rtmidi2==1.4.1
runs==1.2.2
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
xmod==1.8.1
yeelight==0.7.16
